requirements.txt
asgiref==3.7.2
atlassian-python-api==3.40.0
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
constantly==15.1.0
cryptography==41.0.3
daphne==4.0.0
deep-translator==1.11.4
Deprecated==1.2.14
Django==4.2.3
django-admin-autocomplete-filter==0.7.1
django-crontab==0.7.1
django-nested-inline==0.4.6
django-split-settings==1.2.0
gitdb==4.0.10
GitPython==3.1.32
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
oauthlib==3.2.2
packaging==23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==23.2.0
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
sqlparse==0.4.4
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0
zope.interface==6.0
